TitanBot Code/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
branca==0.7.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
deptry==0.16.1
distro==1.9.0
docopt==0.6.2
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.3
fastjsonschema==2.19.1
fiona==1.9.6
folium==0.16.0
frozenlist==1.4.1
geopandas==0.14.4
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.2.3
langchain-cli==0.0.24
langchain-community==0.2.4
langchain-core==0.2.5
langchain-experimental==0.0.60
langchain-nvidia-ai-endpoints==0.1.2
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langserve==0.2.2
langsmith==0.1.77
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numpy==1.26.4
openai==1.34.0
orjson==3.10.4
osmnx==1.9.4
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip-review==1.3.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
streamlit_folium==0.20.1
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.9.4
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.23.2
wcwidth==0.2.13
webencodings==0.5.1
xyzservices==2024.6.0
yarg==0.1.9
yarl==1.9.4
